+++ src/main/java/sudols/ecopercent/dto/item/CreateItemRequest.java
@@ -7,7 +7,7 @@ import sudols.ecopercent.domain.User;
 import java.util.Date;
 
 @Data
-public class RequestPostItemDto {
+public class CreateItemRequest {
     private Long userId;
 
     // TODO: image 자료형 뭐로 해야하는지?
@@ -22,8 +22,6 @@ public class RequestPostItemDto {
     private String brand;
 
     private Integer price;
-
-    private Long goalUsageCount;
 
     private Date purchaseDate;
